refactor: 도메인 모델 패키지 구조 개선 (entity, auth, response 분리)
