+++ booker-server/src/main/java/com/bookerapp/core/domain/model/auth/JwtClaims.java
@@ -1,4 +1,4 @@
-package com.bookerapp.core.domain.model;
+package com.bookerapp.core.domain.model.auth;
 
 public final class JwtClaims {
     private JwtClaims() {}
