+++ booker-server/src/main/java/com/bookerapp/core/domain/model/auth/Role.java
@@ -1,4 +1,4 @@
-package com.bookerapp.core.domain.model;
+package com.bookerapp.core.domain.model.auth;
 
 public enum Role {
     USER("USER"),
